sites/akvo-flow-web-api/requirements.txt
pandas==0.24.2
Flask==1.1.1
Flask-AppBuilder==2.1.4
Flask-Babel==0.12.2
Flask-BasicAuth==0.2.0
Flask-Caching==1.7.2
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.18.2
Flask-Login==0.4.1
Flask-HTTPAuth==4.0.0
Flask-Migrate==2.5.2
Flask-OpenID==1.2.5
xmltodict==0.11.0
requests==2.21.0
lxml==4.3.0
Werkzeug==0.16.1
psycopg2-binary==2.8.5
SQLAlchemy==1.3.17
alembic==1.4.2
